LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.drawing.style.nonfinite.spacing-expected.txt

2d.text.drawing.style.nonfinite.spacing
Testing letter spacing and word spacing with nonfinite inputs
Actual output:

FAIL Testing letter spacing and word spacing with nonfinite inputs assert_equals: ctx.letterSpacing === '0px' (got [undefined], expected 0px[string]) expected (string) "0px" but got (undefined) undefined
